improved representation of collections' lists
we want a collection's lists to comprise a set of unique ID strings. To achieve this, while using the Couchbase subcodcument API to avoid unnecessary copying of data, we represent the set as a JSON object, storing the ID strings as the keys of the object. The values are arbitrary; for now, they are always null.