switched back to using the delectus-api subproject for the API server.
It was misguided to combine the API server with the webapp; they need to be separate for security reasons, and authentication to the user-facing Delectus app should work differently from authetication to the admin-and-dev-only API server, which should in any case never be visible to anyone outside out private network.

Finally, I learned how to enable nrepl sessions from within the compojure-api server process, so I don't need the webapp superstructure to support that.