found and fixed a library bug:
I was relying on the system version of SQLite3. That's fine on Catalina, which I was using for development, but fails on some earlier versions.

The issues was that Delectus uses window functions (specifically, row_number over (partition by ...)), and those were not supported in SQLite until version 3.25; macOS Mojave used SQLite 3.24.

The solution is to ship a newer version of SQLite built into the application.  This commit adds a Homebrew build of libsqlite3 version 3.28.0, and modifies the loading code to load it instead of the default system version.

TODO: modify the build script to build the library into the built app, and the library-loading code to load it from the app bundle.