separating the edit log from the listfile. Henceforth the editlog will be a text interchange file in jsonl format, in which edits are represented as JSON object, one per line, in causal order. The list file will contain SQLite tables for the list metadata (carried in the initial heaer object in the editlog), listname ops, comment ops, columns ops, and item ops.
Delectus will serialize the data in a list file by writing its metadata to a header object, and then writing the stored ops to the jsonl log in causal/timestamp order. It can read an editlog to reconstruct a listfile or to merge ops from a remote listfile into the local one.